Restructure the quota apis in preparation for paid for storage
