+++ src/peergos/server/net/StorageHandler.java
@@ -55,7 +55,7 @@ public class StorageHandler implements HttpHandler {
                     break;
                 case "request":
                     byte[] signedReq = ArrayOps.hexToBytes(last.apply("req"));
-                    boolean res = spaceUsage.requestSpace(owner, signedReq).join();
+                    boolean res = spaceUsage.requestQuota(owner, signedReq).join();
                     result = new CborObject.CborBoolean(res);
                     break;
                 default:
